+++ workspace/CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 * `workspace.run`: include the task duration in milliseconds in the exported `json`.
 
+* `workspace.list`: support `--base` and `--head` options.
+
 * `workspace.test.coverage`: allow filtering by package path.
 
 ## [v0.2.1](https://github.com/sportradar/elixir-workspace/tree/workspace/v0.2.0) (2025-03-14)